# Add a basic alias analysis with redundant-load elim and store-to-load fowarding opts.