--- a/cranelift/codegen/src/context.rs
+++ b/cranelift/codegen/src/context.rs
@@ -9,6 +9,7 @@
 //! contexts concurrently. Typically, you would have one context per compilation thread and only a
 //! single ISA instance.
 
+use crate::alias_analysis::AliasAnalysis;
 use crate::binemit::CodeInfo;
 use crate::dce::do_dce;
 use crate::dominator_tree::DominatorTree;
@@ -161,6 +162,11 @@ impl Context {
         }
 
         self.remove_constant_phis(isa)?;
+
+        if opt_level != OptLevel::None {
+            self.replace_redundant_loads()?;
+            self.simple_gvn(isa)?;
+        }
 
         let result = isa.compile_function(&self.func, self.want_disasm)?;
         let info = result.code_info();
@@ -339,6 +345,17 @@ impl Context {
     {
         eliminate_unreachable_code(&mut self.func, &mut self.cfg, &self.domtree);
         self.verify_if(fisa)
+    }
+
+    /// Replace all redundant loads with the known values in
+    /// memory. These are loads whose values were already loaded by
+    /// other loads earlier, as well as loads whose values were stored
+    /// by a store instruction to the same instruction (so-called
+    /// "store-to-load forwarding").
+    pub fn replace_redundant_loads(&mut self) -> CodegenResult<()> {
+        let mut analysis = AliasAnalysis::new(&mut self.func, &self.domtree);
+        analysis.compute_and_update_aliases();
+        Ok(())
     }
 
     /// Harvest candidate left-hand sides for superoptimization with Souper.