--- a/cranelift/codegen/src/inst_predicates.rs
+++ b/cranelift/codegen/src/inst_predicates.rs
@@ -1,6 +1,10 @@
 //! Instruction predicates/properties, shared by various analyses.
 
-use crate::ir::{DataFlowGraph, Function, Inst, InstructionData, Opcode};
+use crate::ir::immediates::Offset32;
+use crate::ir::instructions::BranchInfo;
+use crate::ir::{
+    Block, DataFlowGraph, Function, Inst, InstructionData, MemFlags, Opcode, Type, Value,
+};
 use crate::machinst::ty_bits;
 use cranelift_entity::EntityRef;
 
@@ -76,5 +80,83 @@ pub fn is_constant_64bit(func: &Function, inst: Inst) -> Option<u64> {
             Some(imm)
         }
         _ => None,
+    }
+}
+
+/// Get the MemFlags from the given instruction, if any.
+pub fn inst_memflags(func: &Function, inst: Inst) -> Option<MemFlags> {
+    let data = &func.dfg[inst];
+    match data {
+        InstructionData::AtomicRmw { flags, .. }
+        | InstructionData::AtomicCas { flags, .. }
+        | InstructionData::Load { flags, .. }
+        | InstructionData::LoadNoOffset { flags, .. }
+        | InstructionData::Store { flags, .. }
+        | InstructionData::StoreNoOffset { flags, .. } => Some(flags.clone()),
+        _ => {
+            debug_assert!(!data.opcode().can_load() && !data.opcode().can_store());
+            None
+        }
+    }
+}
+
+/// Get the address, offset, and access type from the given instruction, if any.
+pub fn inst_addr_offset_type(func: &Function, inst: Inst) -> Option<(Value, Offset32, Type)> {
+    let data = &func.dfg[inst];
+    match data {
+        InstructionData::Load { arg, offset, .. } => {
+            let ty = func.dfg.value_type(func.dfg.inst_results(inst)[0]);
+            Some((*arg, *offset, ty))
+        }
+        InstructionData::LoadNoOffset { arg, .. } => {
+            let ty = func.dfg.value_type(func.dfg.inst_results(inst)[0]);
+            Some((*arg, 0.into(), ty))
+        }
+        InstructionData::Store { args, offset, .. } => {
+            let ty = func.dfg.value_type(args[0]);
+            Some((args[1], *offset, ty))
+        }
+        InstructionData::StoreNoOffset { args, .. } => {
+            let ty = func.dfg.value_type(args[0]);
+            Some((args[1], 0.into(), ty))
+        }
+        _ => None,
+    }
+}
+
+/// Get the store data, if any, from an instruction.
+pub fn inst_store_data(func: &Function, inst: Inst) -> Option<Value> {
+    let data = &func.dfg[inst];
+    match data {
+        InstructionData::Store { args, .. } | InstructionData::StoreNoOffset { args, .. } => {
+            Some(args[0])
+        }
+        _ => None,
+    }
+}
+
+/// Visit all successors of a block with a given visitor closure.
+pub(crate) fn visit_block_succs<F: FnMut(Inst, Block)>(f: &Function, block: Block, mut visit: F) {
+    for inst in f.layout.block_likely_branches(block) {
+        if f.dfg[inst].opcode().is_branch() {
+            visit_branch_targets(f, inst, &mut visit);
+        }
+    }
+}
+
+fn visit_branch_targets<F: FnMut(Inst, Block)>(f: &Function, inst: Inst, visit: &mut F) {
+    match f.dfg[inst].analyze_branch(&f.dfg.value_lists) {
+        BranchInfo::NotABranch => {}
+        BranchInfo::SingleDest(dest, _) => {
+            visit(inst, dest);
+        }
+        BranchInfo::Table(table, maybe_dest) => {
+            if let Some(dest) = maybe_dest {
+                visit(inst, dest);
+            }
+            for &dest in f.jump_tables[table].as_slice() {
+                visit(inst, dest);
+            }
+        }
     }
 }