… cranelift/codegen/src/ir/memflags.rs
… cranelift/codegen/src/lib.rs
… cranelift/codegen/src/machinst/blockorder.rs
… cranelift/codegen/src/machinst/lower.rs
… cranelift/filetests/filetests/alias/categories.clif
… cranelift/filetests/filetests/alias/extends.clif
… cranelift/filetests/filetests/alias/fence.clif
… cranelift/filetests/filetests/alias/multiple-blocks.clif
… cranelift/filetests/filetests/alias/partial-redundancy.clif
… cranelift/filetests/filetests/alias/simple-alias.clif
… cranelift/filetests/src/lib.rs
… cranelift/filetests/src/test_alias_analysis.rs
… cranelift/wasm/src/code_translator.rs
… crates/cranelift/src/func_environ.rs