@bjorn3 2022-05-19
By the way what is the time complexity of this optinization pass?