@cfallin 2022-05-19
@bjorn3 I may be misreading your benchmark result, but:

```
Benchmark 1: ./raytracer_cg_clif_release_alias_analysis
  Time (mean ± σ):      5.196 s ±  0.012 s    [User: 5.191 s, System: 0.004 s]

Benchmark 2: ./raytracer_cg_clif_release_main
  Time (mean ± σ):      5.224 s ±  0.024 s    [User: 5.218 s, System: 0.005 s]
```

looks like `alias_analysis` ran a bit faster than `main`? Then `'./raytracer_cg_clif_release_alias_analysis' ran 1.01 ± 0.01 times faster than './raytracer_cg_clif_release_main'` seems to confirm -- +1%, though maybe within a margin of error.