@bjorn3 2022-05-19
> looks like alias_analysis ran a bit faster than main?

Right, my bad.