@cfallin 2022-05-20
@bjorn3 the alias analysis / redundant-load pass runs after legalization, so it should see plain loads/stores rather than `stack_load`/`stack_store`, I think.