@pepyakin 2022-06-18
Sure! 

[wasmibox.wasm.gz](https://github.com/bytecodealliance/wasmtime/files/8933348/wasmibox.wasm.gz)